dissync-kalman-report: plot input and filtered signal

+++ dissync-kalman-py/src/lib.rs
@@ -1,8 +1,8 @@
 use pyo3::prelude::*;
 
 #[pyfunction]
-fn filter(a: Vec<u32>) -> PyResult<Vec<u32>> {
-    let mut b = vec![0; a.len()];
+fn filter(a: Vec<f32>) -> PyResult<Vec<f32>> {
+    let mut b = vec![0.; a.len()];
     dissync_kalman::filter(a.as_slice(), b.as_mut_slice());
     Ok(b)
 }
