+++ dissync-kalman/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn filter(a: &[u32], b: &mut [u32]) {
+pub fn filter(a: &[f32], b: &mut [f32]) {
     b.clone_from_slice(a);
 }
